Fix : 나눗셈 연산 결과 오류 해결
형변환을 해주지않아 나눗셈의 연산식이 int형으로 나오는 문제가 발생하여
double형으로 형변환을 시켜주었다.

+++ src/main/java/com/example/calculator/calculator1/App1.java
@@ -41,7 +41,7 @@ public class App1 {
                     break;
                 case '/':
                     if (input2 != 0) {
-                        System.out.println("연산 결과:" + (input1 / input2));
+                        System.out.println("연산 결과:" + ((double) input1 / input2));
                         break;
                     }
                 default:
